debundle/lowering: extract chunk_renames
BAZEL_TEST_INVOCATIONS=none: file move.
